[dagster-airlift] Unified complete code directory
