+++ examples/experimental/dagster-airlift/examples/tutorial-example/tutorial_example/completed_code_examples/dagster_stages/migrate.py
@@ -1,3 +1,4 @@
+# type: ignore
 # start_migrate
 import os
 from pathlib import Path
