+++ examples/experimental/dagster-airlift/examples/tutorial-example/tutorial_example/completed_code_examples/dagster_stages/migrate_dag_level.py
@@ -1,28 +1,25 @@
+# type: ignore
 import os
 from pathlib import Path
 
 from dagster import (
-    AssetCheckResult,
-    AssetCheckSeverity,
     AssetExecutionContext,
-    AssetKey,
     AssetsDefinition,
     AssetSpec,
     Definitions,
-    asset_check,
     materialize,
     multi_asset,
 )
 from dagster_airlift.core import (
     AirflowInstance,
     BasicAuthBackend,
-    assets_with_task_mappings,
+    assets_with_dag_mappings,
     build_defs_from_airflow_instance,
 )
 from dagster_dbt import DbtCliResource, DbtProject, dbt_assets
 
 # Code also invoked from Airflow
-from tutorial_example.shared.complete.export_duckdb_to_csv import (
+from tutorial_example.complete.export_duckdb_to_csv import (
     ExportDuckDbToCsvArgs,
     export_duckdb_to_csv,
 )
@@ -63,10 +60,9 @@ def dbt_project_assets(context: AssetExecutionContext, dbt: DbtCliResource):
     yield from dbt.cli(["build"], context=context).stream()
 
 
-mapped_assets = assets_with_task_mappings(
-    dag_id="rebuild_customers_list",
-    task_mappings={
-        "load_raw_customers": [
+mapped_assets = assets_with_dag_mappings(
+    dag_mappings={
+        "rebuild_customers_list": [
             load_csv_to_duckdb_asset(
                 AssetSpec(key=["raw_data", "raw_customers"]),
                 LoadCsvToDuckDbArgs(
@@ -77,12 +73,8 @@ mapped_assets = assets_with_task_mappings(
                     duckdb_schema="raw_data",
                     duckdb_database_name="jaffle_shop",
                 ),
-            )
-        ],
-        "build_dbt_models":
-        # load rich set of assets from dbt project
-        [dbt_project_assets],
-        "export_customers": [
+            ),
+            dbt_project_assets,
             export_duckdb_to_csv_defs(
                 AssetSpec(key="customers_csv", deps=["customers"]),
                 ExportDuckDbToCsvArgs(
@@ -91,32 +83,10 @@ mapped_assets = assets_with_task_mappings(
                     duckdb_path=WAREHOUSE_PATH,
                     duckdb_database_name="jaffle_shop",
                 ),
-            )
+            ),
         ],
     },
 )
-
-
-@asset_check(asset=AssetKey(["customers_csv"]))
-def validate_exported_csv() -> AssetCheckResult:
-    csv_path = CUSTOMERS_CSV_PATH
-
-    if not csv_path.exists():
-        return AssetCheckResult(passed=False, description=f"Export CSV {csv_path} does not exist")
-
-    rows = len(csv_path.read_text().split("\n"))
-    if rows < 2:
-        return AssetCheckResult(
-            passed=False,
-            description=f"Export CSV {csv_path} is empty",
-            severity=AssetCheckSeverity.WARN,
-        )
-
-    return AssetCheckResult(
-        passed=True,
-        description=f"Export CSV {csv_path} exists",
-        metadata={"rows": rows},
-    )
 
 
 defs = build_defs_from_airflow_instance(
@@ -130,7 +100,6 @@ defs = build_defs_from_airflow_instance(
     ),
     defs=Definitions(
         assets=mapped_assets,
-        asset_checks=[validate_exported_csv],
         resources={"dbt": DbtCliResource(project_dir=dbt_project_path())},
     ),
 )
