+++ examples/experimental/dagster-airlift/examples/tutorial-example/tutorial_example/completed_code_examples/dagster_stages/migrate_with_check.py
@@ -1,18 +1,23 @@
+# type: ignore
 import os
 from pathlib import Path
 
 from dagster import (
+    AssetCheckResult,
+    AssetCheckSeverity,
     AssetExecutionContext,
+    AssetKey,
     AssetsDefinition,
     AssetSpec,
     Definitions,
+    asset_check,
     materialize,
     multi_asset,
 )
 from dagster_airlift.core import (
     AirflowInstance,
     BasicAuthBackend,
-    assets_with_dag_mappings,
+    assets_with_task_mappings,
     build_defs_from_airflow_instance,
 )
 from dagster_dbt import DbtCliResource, DbtProject, dbt_assets
@@ -59,9 +64,10 @@ def dbt_project_assets(context: AssetExecutionContext, dbt: DbtCliResource):
     yield from dbt.cli(["build"], context=context).stream()
 
 
-mapped_assets = assets_with_dag_mappings(
-    dag_mappings={
-        "rebuild_customers_list": [
+mapped_assets = assets_with_task_mappings(
+    dag_id="rebuild_customers_list",
+    task_mappings={
+        "load_raw_customers": [
             load_csv_to_duckdb_asset(
                 AssetSpec(key=["raw_data", "raw_customers"]),
                 LoadCsvToDuckDbArgs(
@@ -72,8 +78,12 @@ mapped_assets = assets_with_dag_mappings(
                     duckdb_schema="raw_data",
                     duckdb_database_name="jaffle_shop",
                 ),
-            ),
-            dbt_project_assets,
+            )
+        ],
+        "build_dbt_models":
+        # load rich set of assets from dbt project
+        [dbt_project_assets],
+        "export_customers": [
             export_duckdb_to_csv_defs(
                 AssetSpec(key="customers_csv", deps=["customers"]),
                 ExportDuckDbToCsvArgs(
@@ -82,10 +92,32 @@ mapped_assets = assets_with_dag_mappings(
                     duckdb_path=WAREHOUSE_PATH,
                     duckdb_database_name="jaffle_shop",
                 ),
-            ),
+            )
         ],
     },
 )
+
+
+@asset_check(asset=AssetKey(["customers_csv"]))
+def validate_exported_csv() -> AssetCheckResult:
+    csv_path = CUSTOMERS_CSV_PATH
+
+    if not csv_path.exists():
+        return AssetCheckResult(passed=False, description=f"Export CSV {csv_path} does not exist")
+
+    rows = len(csv_path.read_text().split("\n"))
+    if rows < 2:
+        return AssetCheckResult(
+            passed=False,
+            description=f"Export CSV {csv_path} is empty",
+            severity=AssetCheckSeverity.WARN,
+        )
+
+    return AssetCheckResult(
+        passed=True,
+        description=f"Export CSV {csv_path} exists",
+        metadata={"rows": rows},
+    )
 
 
 defs = build_defs_from_airflow_instance(
@@ -99,6 +131,7 @@ defs = build_defs_from_airflow_instance(
     ),
     defs=Definitions(
         assets=mapped_assets,
+        asset_checks=[validate_exported_csv],
         resources={"dbt": DbtCliResource(project_dir=dbt_project_path())},
     ),
 )
