+++ examples/experimental/dagster-airlift/examples/tutorial-example/tutorial_example/completed_code_examples/dagster_stages/observe_with_partitions.py
