+++ examples/experimental/dagster-airlift/examples/tutorial-example/tutorial_example/completed_code_examples/dagster_stages/partial_migrate.py
@@ -1,3 +1,4 @@
+# type: ignore
 import os
 from pathlib import Path
