+++ examples/experimental/dagster-airlift/examples/tutorial-example/tutorial_example/completed_code_examples/shared_code/load_csv_to_duckdb.py
